Refactor common delegate tree transformation code
